backend/requirements.txt
asgiref
Django
django-ckeditor
django-cors-headers
django-js-asset
django-webpack-loader
djangorestframework
pillow
psycopg2-binary
python-dotenv
sqlparse
typing_extensions
tzdata
